Added user profile page (#9)
(#9) - 9 commit
MainController was divided into 2 controllers for convenience, also index controller was changed, now User is passed to it, to be able to access the page available only to this user, the profile page. The profile controller was added, there were also some code improvements, in particular the logic of the changeUserRole method was changed.